+++ seahorn/jobs/storage_ipc_indirect_handlers/CMakeLists.txt
@@ -4,7 +4,7 @@ add_executable(storage_ipc_indirect_handlers
   main.c)
 target_include_directories(storage_ipc_indirect_handlers 
   PRIVATE ${TRUSTY_ROOT}/trusty/user/app/storage)
-sea_overlink_libraries(storage_ipc_indirect_handlers trusty_msg.ir)
+sea_overlink_libraries(storage_ipc_indirect_handlers trusty_api_override.ir)
 sea_overlink_libraries(storage_ipc_indirect_handlers realloc_override.ir)
 sea_overlink_libraries(storage_ipc_indirect_handlers fprintf_override.ir)
 sea_overlink_libraries(storage_ipc_indirect_handlers ipc_loop_override.ir)
