+++ seahorn/jobs/storage_ipc_port_create_destroy/CMakeLists.txt
@@ -5,7 +5,7 @@ add_executable(storage_ipc_port_create_destroy
 target_include_directories(storage_ipc_port_create_destroy 
   PRIVATE ${TRUSTY_ROOT}/trusty/user/app/storage)
 
-sea_overlink_libraries(storage_ipc_port_create_destroy trusty_msg.ir)
+sea_overlink_libraries(storage_ipc_port_create_destroy trusty_api_override.ir)
 sea_overlink_libraries(storage_ipc_port_create_destroy realloc_override.ir)
 sea_overlink_libraries(storage_ipc_msg_buffer fprintf_override.ir)
 sea_attach_bc_link(storage_ipc_port_create_destroy)
